fix integrity meta initializer
