+++ src/main/java/org/dependencytrack/event/IntegrityMetaInitializerEvent.java
@@ -1,0 +1,6 @@
+package org.dependencytrack.event;
+
+import alpine.event.framework.Event;
+
+public class IntegrityMetaInitializerEvent implements Event {
+}
