+++ src/main/java/org/dependencytrack/tasks/LockName.java
@@ -9,5 +9,5 @@ public enum LockName {
     WORKFLOW_STEP_CLEANUP_TASK_LOCK,
     PORTFOLIO_REPO_META_ANALYSIS_TASK_LOCK,
     PORTFOLIO_VULN_ANALYSIS_TASK_LOCK,
-    INTEGRITY_META_INITIALIZER_TASK_LOCK
+    INTEGRITY_META_INITIALIZER_LOCK
 }
