+++ src/main/resources/application.properties
@@ -456,8 +456,8 @@ task.portfolio.repoMetaAnalysis.lockAtMostForInMillis=900000
 task.portfolio.repoMetaAnalysis.lockAtLeastForInMillis=90000
 task.portfolio.vulnAnalysis.lockAtMostForInMillis=900000
 task.portfolio.vulnAnalysis.lockAtLeastForInMillis=90000
-task.integrity.integrityMetaInitializer.lockAtMostForInMillis=900000
-task.integrity.integrityMetaInitializer.lockAtLeastForInMillis=90000
+integrityMetaInitializer.lockAtMostForInMillis=900000
+integrityMetaInitializer.lockAtLeastForInMillis=90000
 
 #schedule task for 10th minute of every hour
 task.cron.metrics.portfolio=10 * * * *
@@ -489,6 +489,8 @@ task.cron.defectdojo.sync=0 2 * * *
 task.cron.kenna.sync=0 2 * * *
 #schedule task every 15 minutes
 task.cron.workflow.state.cleanup=*/15 * * * *
+#schedule task at 0 min past every 12th hr
+task.cron.integrityInitializer=0 */12 * * *
 
 # Optional
 # Defines the number of write operations to perform during BOM processing before changes are flushed to the database.
