+++ src/main/webapp/WEB-INF/web.xml
@@ -57,7 +57,7 @@
         <listener-class>org.dependencytrack.event.kafka.KafkaStreamsInitializer</listener-class>
     </listener>
     <listener>
-        <listener-class>org.dependencytrack.event.IntegrityMetaInitializer</listener-class>
+        <listener-class>org.dependencytrack.event.PurlMigrator</listener-class>
     </listener>
 
     <filter>
