+++ src/test/java/org/dependencytrack/persistence/IntegrityMetaQueryManagerPostgresTest.java
@@ -1,8 +1,10 @@
 package org.dependencytrack.persistence;
 
 import org.dependencytrack.AbstractPostgresEnabledTest;
+import org.dependencytrack.model.Component;
 import org.dependencytrack.model.FetchStatus;
 import org.dependencytrack.model.IntegrityMetaComponent;
+import org.dependencytrack.model.Project;
 import org.junit.Test;
 
 import java.util.Date;
@@ -25,5 +27,38 @@ public class IntegrityMetaQueryManagerPostgresTest extends AbstractPostgresEnabl
         integrityMeta2.setStatus(FetchStatus.IN_PROGRESS);
         integrityMeta2.setLastFetch(new Date());
         assertThat(qm.getIntegrityMetaComponentCount()).isEqualTo(1);
+    }
+
+    @Test
+    public void testSynchronizeIntegrityMetaComponent() {
+        final var project = new Project();
+        project.setName("acme-app");
+        project.setVersion("1.0.0");
+        qm.persist(project);
+
+        final var component = new Component();
+        component.setProject(project);
+        component.setPurl("pkg:maven/acme/example@1.0.0?type=jar");
+        component.setName("acme-lib");
+
+        // without any component in database
+        qm.synchronizeIntegrityMetaComponent();
+        assertThat(qm.getIntegrityMetaComponent(component.getPurl().toString())).isNull();
+
+        // with existing component in database
+        qm.persist(component);
+        qm.synchronizeIntegrityMetaComponent();
+        assertThat(qm.getIntegrityMetaComponent(component.getPurl().toString())).satisfies(
+                meta -> {
+                    assertThat(meta.getStatus()).isNull();
+                    assertThat(meta.getPurl()).isEqualTo("pkg:maven/acme/example@1.0.0?type=jar");
+                    assertThat(meta.getId()).isEqualTo(1L);
+                    assertThat(meta.getMd5()).isNull();
+                    assertThat(meta.getSha1()).isNull();
+                    assertThat(meta.getSha256()).isNull();
+                    assertThat(meta.getLastFetch()).isNull();
+                    assertThat(meta.getPublishedAt()).isNull();
+                }
+        );
     }
 }
