+++ src/test/java/org/dependencytrack/persistence/IntegrityMetaQueryManagerTest.java
@@ -56,39 +56,6 @@ public class IntegrityMetaQueryManagerTest extends PersistenceCapableTest {
     }
 
     @Test
-    public void testSynchronizeIntegrityMetaComponent() {
-        final var project = new Project();
-        project.setName("acme-app");
-        project.setVersion("1.0.0");
-        qm.persist(project);
-
-        final var component = new Component();
-        component.setProject(project);
-        component.setPurl("pkg:maven/acme/example@1.0.0?type=jar");
-        component.setName("acme-lib");
-
-        // without any component in database
-        qm.synchronizeIntegrityMetaComponent();
-        assertThat(qm.getIntegrityMetaComponent(component.getPurl().toString())).isNull();
-
-        // with existing component in database
-        qm.persist(component);
-        qm.synchronizeIntegrityMetaComponent();
-        assertThat(qm.getIntegrityMetaComponent(component.getPurl().toString())).satisfies(
-                meta -> {
-                    assertThat(meta.getStatus()).isNull();
-                    assertThat(meta.getPurl()).isEqualTo("pkg:maven/acme/example@1.0.0?type=jar");
-                    assertThat(meta.getId()).isEqualTo(1L);
-                    assertThat(meta.getMd5()).isNull();
-                    assertThat(meta.getSha1()).isNull();
-                    assertThat(meta.getSha256()).isNull();
-                    assertThat(meta.getLastFetch()).isNull();
-                    assertThat(meta.getPublishedAt()).isNull();
-                }
-        );
-    }
-
-    @Test
     public void testGetIntegrityMetaComponentCount() {
         var integrityMeta = new IntegrityMetaComponent();
         integrityMeta.setPurl("pkg:maven/acme/example@1.0.0?type=jar");
